requirements.txt
customtkinter
python-dotenv
CTkMessagebox